MIPS: Set USE_MUL to 0 by default (for now).
Setting it to 1 breaks on Loongson CPUs which are the most common
desktop MIPS CPUs supported by e.g. Debian.

This change can be reverted when more investigation is done to find
out why exactly this breaks.

Thanks to Jo Shields (@directhex) for lots of testing and debugging
and spotting this define that turned out to be the culprit.
(cherry picked from commit 0e67acba31c39e37b1bc55afb28d157c36895c23)