[runtime] Apply interruption deferring also to global interruption count
When aborting a thread we set the thread's interruption request and we also increment the global interruption request. If we were trying to abort a thread when it is in an abort protected region, it will just set an interruption deferred and not do anything else. When the thread exited the abort protected region, it was missing the abort request since we didn't increment hte global interruption request.

In order to fix this, while a thread (that is to be aborted) is in a protected block, we keep the global interruption count unchanged, and increment it when exiting the block, if we have an interruption pending. This means that if we are entering an abort protected block and we are deferring an existing interrupt request we also need to decrement the global var.