[runtime] Rework how the native-to-managed eh callback is called. (#5837)
* [runtime] Rework how the native-to-managed eh callback is called.

Instead of catching all exceptions in the native-to-managed wrapper and calling the callback from the catch clause, call it directly from the EH code. The
previous approach caught all exceptions, which messed up debugging, since the debugger wouldn't break on uncaught exceptions raised from native-to-managed
callbacks.

* [coop] Enable tests bug-58782-plain-throw and bug-58782-capture-and-throw for PRs

* [mini] Installed native-to-managed exception handling callback from second pass.

The constraints are:
1. We must allow managed exceptions to be caught and finally blocks must run
2. We must not call the unmanaged exception handler if the first unwinding pass
   detects that there is a native-to-managed wrapper and a callback is
   installed that will handle managed exceptions (by unwinding the native code
   and either re-raising the managed exception or otherwise dealing with it).

So we change the return value of handle_exception_first_pass to be:
- MONO_FIRST_PASS_HANDLED - caught by a managed frame (with a matching filter if one was present).
- MONO_FIRST_PASS_CALLBACK_TO_NATIVE - there's a native-to-managed wrapper and a ftnptr_eh_callback is installed.
- MONO_FIRST_PASS_UNHANDLED - first pass unwound all the way and there was no wrapper or there was no callback installed.

In mono_handle_exception_internal, we treat MONO_FIRST_PASS_CALLBACK_TO_NATIVE
as handled and unwind the managed stack, invoking finally clauses as we go,
until we reach the managed-to-native frame, at which point we invoke the callback.

* [test] Test for mono_install_ftnptr_eh_callback

* [test] AOT-friendly test for mono_install_ftnptr_eh_callback