System.Diagnostics.Process improvements for AIX (#9180)
* Fix Process.ProcessName on AIX

AIX has a procfs with raw structures congruent with Solaris. For i,
which lacks procfs, fall back to a libc call.

This doesn't solve the Process weirdness that bungs up the test
suite, but I hope it's a start.

* Bye bye, annoying AIX setschedparam warning

AIX (but weirdly, not i) complains every time you do this. Squelch
the warning on AIX and avoid line noise.

* Remove more assumptions of /proc on AIX

	* should probably always use getprocs64 for w32process,
	  but looks like we could just use proclib on the fallback?

	* there's some more cases relating to resource usage stat,
	  will have to investigate those cases further

* Basic process modules support on AIX (but not i)

Not aware of a PASE compatible API, but this will work on AIX. Not
sure about .ModuleName versus .FileName, but it seems to fill it in...