Beginning of interpreter on PowerPC again
This is useful for diagnosing issues with the JIT and running under
more restricted enviroments.

The interpreter starts to run, but will fail on a managed-to-native
transition. It must be decided to have JITted frames assist or use
the C switch trampolines like WebAssembly. It's more complex, but
worthwhile to do the PPC JIT work.

Changes include:

	* Fix context to use bp/sp always on sc_sp. The getter and
	  setter for BP was different, which confused exception
	  handling. This is apparently subtlely weird due to PPC
	  sigcontext. Perhaps we can do what z does and set both
	  r1 (stack) and r31 (frame).

	* Disable fast icalls on PPC due to endianness.

	* Enable interpreter on PPC in headers.

	* Implement IP adjustment for PPC, needed for interp.

	* Stub for managed-to-native transition.