Significantly decrease nonrecursive interpreter memory use.
This is a more correct, but still flawed, alternative
to https://github.com/mono/mono/pull/18756
and should go in before https://github.com/mono/mono/pull/18734,
though I can add it to it.

So, this PR would appear to use more memory, not less.
And sometimes it does.
But it will tend to use far less memory.

The problem is that the interpreter implements a custom
heap, and does so drastically favoring speed of allocation
over density.

Any allocation, big or small, that does not fit in the current
chunk, forces a new chunk to be allocated, and leaves unused
the tail of the prior chunk. Also chunks are never freed
until the entire thread is freed.

This larger 64K (or whatever is chosen) will still behave poorly
when straddling occurs, but it will much less often.

There are other approaches to consider:

 1. Only allocate one large chunk, that correlates with native
   stack size. Like 512k or 1mb or 8mb or whatever.
   Maybe, maybe not be willing to allocate more than one.

 2. The "split" technique, which does perfect fit frame
   allocation in the usual fixed size or alloca way.

 3. An as yet unimplemented idea:
   a. Multiple alloca in same frame are adjacent.
   b. Manage one growable chunk per frame via multiple alloca.
   c. Chunk growth is alloca.
   d. Chunk reclaim is only via return, like alloca.
   e. Free is setting the pointer back.
   f. Allocate is "bump", possibly with alloca for growth.
   g. Dynamically detect stack growth direction and manage accordingly.

   This should be tried later perhaps.
   The notion of adjacent repeated alloca is a little gross, but
   is used successfully in practise. Consider that alloca itself is not portable.

  4. Maybe be willing to g_free chunks in frame_stack_pop.
     This would suffer when going back and forth across boundaries,
     repeated g_free/g_malloc cycles.