[metadata-update] compute_unaligned_sizes (String heap only) and start_encmap
compute_unaligned_sizes find the size of the heaps in the base image before
they were rounded up to a multiple of 4.  The dmeta image heaps are appended to
the unaligned sizes.

The EncRecs table is adapted form CMiniMdRW::StartENCMap
https://github.com/dotnet/runtime/blob/4f9ae42d861fcb4be2fcd5d3d55d5f227d30e723/src/coreclr/src/md/enc/metamodelenc.cpp#L215
it's a way to quickly go from a table index to its first entry in the ENCMap